Add basic UI Add UI logic (WIP)

+++ src/js/secondary.js
@@ -21,7 +21,7 @@ export default class personsAge {
         this.futureAge = futureAge;
     }
     earthAge() {
-        return parseInt(this.age / 1)  + " Years Old";
+        return parseInt(this.age / 1) + " Years Old";
     }
     mercuryAge() {
         return parseFloat(this.age / 0.24) + " Years Old";
@@ -33,38 +33,38 @@ export default class personsAge {
         return parseFloat(this.age / 1.88) + " Years Old";
     }
     jupiterAge() {
-        return parseFloat(this.age / 11.86) + " Years Old"; 
+        return parseFloat(this.age / 11.86) + " Years Old";
     }
     //Methods used to determine difference in years
     earthPassedYears() {
-        return parseInt(this.secondAge / 1)  - parseInt(this.age / 1) + " Years have passed";
+        return parseInt(this.secondAge / 1) - parseInt(this.age / 1) + " Years have passed";
     }
     mercuryPassedYears() {
-        return parseFloat(this.secondAge / 0.24) - parseInt(this.age / 0.24) + " Years have passed";
+        return parseFloat(this.secondAge / 0.24) - parseFloat(this.age / 0.24) + " Years have passed";
     }
     venusPassedYears() {
-        return parseFloat(this.secondAge / 0.62) - parseInt(this.age / 0.62) + " Years have passed";
+        return parseFloat(this.secondAge / 0.62) - parseFloat(this.age / 0.62) + " Years have passed";
     }
     marsPassedYears() {
-        return parseFloat(this.secondAge / 1.88) - parseInt(this.age / 1.88) + " Years have passed";
+        return parseFloat(this.secondAge / 1.88) - parseFloat(this.age / 1.88) + " Years have passed";
     }
     jupiterPassedYears() {
-        return parseFloat(this.secondAge / 11.86) - parseInt(this.age / 11.86) + " Years have passed"; 
+        return parseFloat(this.secondAge / 11.86) - parseFloat(this.age / 11.86) + " Years have passed";
     }
     //Methods used to determine difference in years for present and future years
     earthFutureYears() {
-        return parseInt(this.futureAge / 1)  - parseInt(this.age / 1) + " Years to go";
+        return parseInt(this.futureAge / 1) - parseInt(this.age / 1) + " Years to go";
     }
     mercuryFutureYears() {
-        return parseFloat(this.futureAge / 0.24) - parseInt(this.age / 0.24) + " Years to go";
+        return parseFloat(this.futureAge / 0.24) - parseFloat(this.age / 0.24) + " Years to go";
     }
     venusFutureYears() {
-        return parseFloat(this.futureAge / 0.62) - parseInt(this.age / 0.62) + " Years to go";
+        return parseFloat(this.futureAge / 0.62) - parseFloat(this.age / 0.62) + " Years to go";
     }
     marsFutureYears() {
-        return parseFloat(this.futureAge / 1.88) - parseInt(this.age / 1.88) + " Years to go";
+        return parseFloat(this.futureAge / 1.88) - parseFloat(this.age / 1.88) + " Years to go";
     }
     jupiterFutureYears() {
-        return parseFloat(this.futureAge / 11.86) - parseInt(this.age / 11.86) + " Years to go"; 
+        return parseFloat(this.futureAge / 11.86) - parseFloat(this.age / 11.86) + " Years to go";
     }
 }
